+++ cogs/actions.py
@@ -1,234 +1,98 @@
 import discord
 from discord.ext import commands
 import aiohttp
-import datetime
+from datetime import datetime
+import typing
 
 class Actions(commands.Cog, description="interact using gifs!"):
     def __init__(self, bot):
         self.bot = bot
 
-    # interactive
-    @commands.hybrid_command(help="hug someone")
-    async def hug(self, ctx, member: discord.Member):
+    class InteractFlags(commands.FlagConverter):
+        action: typing.Literal[
+            "hug", "pat", "slap", "kiss", "poke", "feed", "bite", 
+            "cuddle", "stare", "tickle", "wave", "highfive", "handhold", 
+            "wink", "handshake", "baka", "kick", "punch", "peck", "shoot", "yeet"
+        ] = commands.flag(description="the action to perform")
+        member: discord.Member = commands.flag(description="the member to interact with")
+
+    class ReactFlags(commands.FlagConverter):
+        action: typing.Literal[
+            "blush", "bored", "cry", "dance", "facepalm", "happy", "laugh",
+            "lurk", "nod", "nom", "nope", "pout", "shrug", "smile", "smug",
+            "think", "thumbsup", "yawn"
+        ] = commands.flag(description="the action to perform")
+
+    @commands.hybrid_command(help="interact with someone!")
+    async def interact(self, ctx, flags: InteractFlags):
+        actions = [
+            "hug", "pat", "slap", "kiss", "poke", "feed", "bite", 
+            "cuddle", "stare", "tickle", "wave", "highfive", "handhold", 
+            "wink", "handshake", "baka", "kick", "punch", "peck", "shoot", "yeet"
+        ]
+        if ctx.author == flags.member:
+            if flags.action in ["hug", "pat", "cuddle", "wave", "handhold", "highfive"]:
+                return await ctx.send("lol loser!")
+            elif flags.action in ["kiss", "peck", "wink", "handshake", "stare", "kick"]:
+                return await ctx.send("uhh how would you even do that to yourself?")
+            elif flags.action in ["baka", "poke", "tickle"]:
+                return await ctx.send("...you're weird")
+            elif flags.action in ["slap", "punch", "shoot", "yeet"]:
+                return await ctx.send("please dont do that! :(") 
+            elif flags.action == "feed":
+                return await ctx.send("you ate some food! yay!\n(for a gif, do `<prefix>react nom`)")
+
+        if flags.action not in actions:
+            return await ctx.send(f"Invalid action! Choose from: ```{', '.join(actions)}```")
+
         async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/hug") as response:
+            async with session.get(f"https://nekos.best/api/v2/{flags.action}") as response:
                 data = await response.json()
+                data = data["results"][0]
                 embed = discord.Embed(
-                    title=f"{ctx.author.name} hugs {member.name}",
                     color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
+                    timestamp=datetime.now()
                 )
+                if flags.action in ["kiss", "punch"]:
+                    embed.title = f"{ctx.author.name} {flags.action}es {flags.member.name}"
+                elif flags.action in ["stare", "wave", "wink"]:
+                    embed.title = f"{ctx.author.name} {flags.action}s at {flags.member.name}"
+                elif flags.action == "baka":
+                    embed.title = f"{ctx.author.name} calls {flags.member.name} a baka"
+                elif flags.action == "handhold":
+                    embed.title = f"{ctx.author.name} holds {flags.member.name}'s hand"
+                else:
+                    embed.title = f"{ctx.author.name} {flags.action}s {flags.member.name}"
+                embed.set_footer(text=f"gif from {data["anime_name"]}")
                 embed.set_image(url=data["url"])
                 await ctx.send(embed=embed)
 
-    @commands.hybrid_command(help="pat someone")
-    async def pat(self, ctx, member: discord.Member):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/pat") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} pats {member.name}",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-
-    @commands.hybrid_command(help="slap someone")
-    async def slap(self, ctx, member: discord.Member):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/slap") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} slaps {member.name}",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-
-    @commands.hybrid_command(help="kiss someone")
-    async def kiss(self, ctx, member: discord.Member):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/kiss") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} kisses {member.name}",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="poke someone")
-    async def poke(self, ctx, member: discord.Member):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/poke") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} pokes {member.name}",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="feed someone")   
-    async def feed(self, ctx, member: discord.Member):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/feed") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} feeds {member.name}",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="nom someone")
-    async def nom(self, ctx, member: discord.Member):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/nom") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} noms {member.name}",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                                
     # reactions
-    @commands.hybrid_command(help="blush")
-    async def blush(self, ctx):
+    @commands.hybrid_command(help="do a reaction!", aliases=["act"])
+    # async def react(self, ctx, action):
+    # show all the actions available when using slash commands
+    async def react(self, ctx, flags: ReactFlags):
+        actions = [
+            "blush", "bored", "cry", "dance", "facepalm", "happy", "laugh",
+            "lurk", "nod", "nom", "nope", "pout", "shrug", "smile", "smug",
+            "think", "thumbsup", "yawn"
+        ]
+
+        if flags.action not in actions:
+            return await ctx.send(f"Invalid action! Choose from: ```{', '.join(actions)}```")
+
         async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/blush") as response:
+            async with session.get(f"https://nekos.best/api/v2/{flags.action}") as response:
                 data = await response.json()
+                data = data["results"][0]
                 embed = discord.Embed(
-                    title=f"{ctx.author.name} blushes",
+                    title=f"{ctx.author.name} {flags.action}s",
                     color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
+                    timestamp=datetime.now(),
                 )
+                embed.set_footer(text=f"gif from {data["anime_name"]}")
                 embed.set_image(url=data["url"])
                 await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="cry")
-    async def cry(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/cry") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} cries",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="laugh")
-    async def laugh(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/laugh") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} laughs",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="pout")
-    async def pout(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/pout") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} pouts",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="smile")
-    async def smile(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/smile") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} smiles",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="smug")
-    async def smug(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/smug") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} smugs",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="stare")
-    async def stare(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/stare") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} stares",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
-    @commands.hybrid_command(help="think")
-    async def think(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/think") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} thinks",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-             
-    @commands.hybrid_command(help="wink")
-    async def wink(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/wink") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} winks",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-               
-    @commands.hybrid_command(help="facepalm")
-    async def facepalm(self, ctx):
-        async with aiohttp.ClientSession() as session:
-            async with session.get("https://api.waifu.pics/sfw/face-palm") as response:
-                data = await response.json()
-                embed = discord.Embed(
-                    title=f"{ctx.author.name} facepalms",
-                    color=ctx.author.color,
-                    timestamp=datetime.datetime.utcnow()
-                )
-                embed.set_image(url=data["url"])
-                await ctx.send(embed=embed)
-                
+
 async def setup(bot):
     await bot.add_cog(Actions(bot))